uses clock_gettime

+++ packet.hh
@@ -3,9 +3,9 @@ using namespace std;
 class Packet {
     public:
         // time packet should be sent
-        int timestamp;
+        uint64_t timestamp;
         string contents;
-        Packet(int i = 0, string j = ""):timestamp(i), contents(j) {};
-        int get_timestamp() const {return timestamp;}
+        Packet(uint64_t i = 0, string j = ""):timestamp(i), contents(j) {};
+        uint64_t get_timestamp() const {return timestamp;}
         string get_content() const {return contents;}
 };
